feat: 02-06작업 내역

+++ src/main/java/com/green/attaparunever2/restaurant/model/SelRestaurantMainRes.java
@@ -19,7 +19,7 @@ public class SelRestaurantMainRes {
     @Schema(title = "식당 주소")
     private String restaurantAddress;
     @Schema(title = "식당 평균 식사 시간")
-    private double avgRestaurant;
+    private int avgRestaurant;
     @Schema(title = "식당 사진 리스트")
     private RestaurantPicAroundSel restaurantAroundPicList;
 }
